extfs: Swap order of member variables
Since we know that sizeof(extfs_superblock) is guaranteed to be 1024,
we should move it to the begining of the data member list. Since 1024
is a power of two, we can be pretty certain, that no padding will be
inserted after fs::extfs::m_primarySuperblock. std::fstream however
makes no guarantees as to what it size is. Thus we would risk the
insertion of padding.